make a very long constant unsigned long long instead of unsigned long. Fix a compiler warning because of an overflow.
git-svn-id: https://svn.lfbs.rwth-aachen.de/svn/mp-mpich/branches/rc-1.5@5223 3ddfa0b1-c7ec-0310-823c-dd91aa7df4cc

+++ mpid/ch_usock/usocktcpcomm.c
@@ -477,7 +477,7 @@ int ResolveAddress(const char *Host,unsigned long *Addr)
   Address = inet_addr(Host);
   
   /* if hostname was not in '.' notation, inet_addr returns -1 */
-  if(Address == ((unsigned long)256*256*256*256-1))
+  if(Address == ((unsigned long long)256*256*256*256-1))
   {
     if ((he=gethostbyname(Host)) != 0)
     {
